% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Fib.R
\name{Fib}
\alias{Fib}
\title{Fibonacci Sequence}
\usage{
Fib(x)
}
\arguments{
\item{x}{A numeric value as the length of fibonacci sequence}
}
\value{
The fibonacci sequence with length x
}
\description{
Takes in any numeric value as the length of fibonacci sequence, and print out the fibonacci sequence.
}

